fix: detail-panel.test.tsx
- brought back detail panel because noticed test was failing
- probably some code relies on specific location, the test passes if we declare routes with /folder, so to avoid any regression I put it back as before
- added folder-list-panel.tsx with similar logic of detail-panel.test.tsx
