+++ src/legacy/views/app/detail-panel.test.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { faker } from '@faker-js/faker';
 import { http, HttpResponse } from 'msw';
 
-import DetailPanel from './detail-panel';
+import { DetailPanel } from './detail-panel';
 import { FOLDERS } from '../../../carbonio-ui-commons/constants/folders';
 import { getSetupServer } from '../../../carbonio-ui-commons/test/jest-setup';
 import { populateFoldersStore } from '../../../carbonio-ui-commons/test/mocks/store/folders';
